requirements.txt
discord.py
mss
opencv-python
numpy